third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/auto-scrollbars-001-expected.txt

This is a testharness.js-based test.
FAIL Initial layout - expecting a scrollbar without overflowing content instead of overflowing content without a scrollbar assert_equals: Layout with a scrollbar means the container query applies expected "50px" but got "100px"
FAIL Same result after a reflow assert_equals: Layout with a scrollbar means the container query applies expected "50px" but got "100px"
Harness: the test ran to completion.
